ref: go-swagger instead of oapi-codegen, update spec, compendium handlers

+++ internal/domain/weapon_class.go
@@ -1,6 +1,6 @@
 package domain
 
 type WeaponClass struct {
-	ID    int8   `json:"id"`
+	ID    int32  `json:"id"`
 	Class string `json:"class"`
 }
